requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Training utilities
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0

# Code analysis
ast-comments>=1.0.0
astunparse>=1.6.3
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Evaluation metrics
sacrebleu>=2.3.1
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
pylint>=2.17.0
black>=23.3.0